requirements.txt
solana==0.30.2
web3==6.11.1
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.4.2
python-multipart==0.0.6
langchain==0.0.330
numpy==1.24.3
pandas==2.0.3
cryptography==41.0.4
python-dotenv==1.0.0
sqlalchemy==2.0.22
